+++ NEWS.md
@@ -1,3 +1,15 @@
+# development
+
+## Breaking
+
+* Instead of `attr(., "JSON")`, the JSON object is now a hidden column (`..JSON`)
+
+## Bug fixes
+
+* Address backwards incompatibilities in `dplyr`
+
+* Remove `tidyjson::bind_rows()` in favor of re-exporting `dplyr::bind_rows()`
+
 # tidyjson 0.2.4
 
 ## New functions